# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OxyGen CMS API
  version: 0.8.0
  summary: OxyGen Central Management Server API
  description: |
    API contract for OxyGen Central Management Server Phase 1.

    Current CMS schema version: 0.19.

    Phase 1 APIs cover first-run setup, local authentication/RBAC, tenant/user/group/role administration,
    manual OxyGen instance enrollment, dashboard background poller status/controls, application logs, grid preferences, application settings,
    Settings system visibility, and Phase 1.5 queue status. Protected routes use bearer authentication plus explicit permissions and Tenant/global/instance scope checks.
servers:
  - url: http://localhost:3000
    description: Local API development server
  - url: https://{host}
    description: Deployed CMS server
    variables:
      host:
        default: cms.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/system/poller:
    get:
      tags: [System]
      summary: Get background polling runner status
      description: Returns the shared background polling runner state used by the dashboard Service card.
      responses:
        '200':
          description: Poller status.
          content:
            application/json:
              schema:
                type: object
                required: [poller]
                properties:
                  poller: { $ref: '#/components/schemas/InstancePollerStatus' }
        '401': { $ref: '#/components/responses/Error' }
  /api/system/queues:
    get:
      tags: [System]
      summary: Get BullMQ queue foundation status
      description: SystemAdmin only. Returns disabled-mode queue definitions when BullMQ is off, or Redis/BullMQ connectivity and queue counts when enabled. Bull Board, when enabled, is mounted separately at the configured admin path behind the same SystemAdmin permission.
      responses:
        '200':
          description: Queue orchestration status.
          content:
            application/json:
              schema:
                type: object
                required: [queues]
                properties:
                  queues: { $ref: '#/components/schemas/QueueStatusSnapshot' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/queue-jobs:
    get:
      tags: [System]
      summary: Get sanitized BullMQ job summaries
      description: SystemAdmin only. Returns native CMS queue visibility with active, ready, scheduled, retained failed, and completed job summaries from named queues. Raw BullMQ payloads are reduced to safe public metadata only; recurring maintenance schedules can be run now without changing enabled/paused state.
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
          description: Maximum number of latest jobs to return across queues.
      responses:
        '200':
          description: Sanitized latest queue jobs.
          content:
            application/json:
              schema:
                type: object
                required: [queueJobs]
                properties:
                  queueJobs: { $ref: '#/components/schemas/QueueJobsSnapshot' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/queue-jobs/{key}/pause:
    post:
      tags: [System]
      summary: Pause a recurring queue schedule
      description: SystemAdmin only. Persists the selected recurring maintenance scheduler as disabled without deleting historical BullMQ job data. Instance scheduler keys use the `instance-check:{instanceId}` form and map to the instance enabled flag.
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
          description: URL-encoded queue schedule key, for example `database-maintenance:purge-logs`.
      responses:
        '200':
          description: Updated queue schedule settings.
          content:
            application/json:
              schema:
                type: object
                required: [queueSchedules]
                properties:
                  queueSchedules: { $ref: '#/components/schemas/QueueScheduleSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/system/queue-jobs/{key}/resume:
    post:
      tags: [System]
      summary: Resume a recurring queue schedule
      description: SystemAdmin only. Persists the selected recurring maintenance scheduler as enabled and reconciles BullMQ scheduler state. Instance scheduler keys use the `instance-check:{instanceId}` form and map to the instance enabled flag.
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
          description: URL-encoded queue schedule key, for example `system-maintenance:check-application-updates`.
      responses:
        '200':
          description: Updated queue schedule settings.
          content:
            application/json:
              schema:
                type: object
                required: [queueSchedules]
                properties:
                  queueSchedules: { $ref: '#/components/schemas/QueueScheduleSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/system/queue-jobs/{key}/run-now:
    post:
      tags: [System]
      summary: Queue a one-off run for a scheduled job
      description: SystemAdmin only. Adds an immediate one-off job for the selected maintenance or instance scheduler without changing the recurring enabled/paused state.
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
          description: URL-encoded queue schedule key, for example `database-maintenance:purge-logs` or `instance-check:{instanceId}`.
      responses:
        '202':
          description: One-off job queued.
          content:
            application/json:
              schema:
                type: object
                required: [queued, key, jobId]
                properties:
                  queued: { type: boolean, enum: [true] }
                  key: { type: string }
                  jobId: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /api/system/update-status:
    get:
      tags: [System]
      summary: Get CMS update orchestration status
      description: SystemAdmin/update-panel endpoint for the non-technical in-place update flow. Returns the current guarded update command status, ordered progress steps, and last in-memory runner result. Runner execution is disabled by default unless deployment configuration explicitly enables it.
      responses:
        '200':
          description: Update orchestration status.
          content:
            application/json:
              schema:
                type: object
                required: [updateStatus]
                properties:
                  updateStatus: { $ref: '#/components/schemas/SystemUpdateStatus' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/update-runner/dry-run:
    post:
      tags: [System]
      summary: Start guarded CMS update dry run
      description: SystemAdmin/settings-management endpoint. Starts the configured host update script in dry-run mode only when `CMS_UPDATE_RUNNER_ENABLED=true`; disabled deployments return a guarded error and never spawn the script.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                targetRef: { type: [string, 'null'], description: Optional release, tag, branch, or commit ref to validate. }
      responses:
        '202':
          description: Dry-run request accepted.
          content:
            application/json:
              schema:
                type: object
                required: [updateStatus]
                properties:
                  updateStatus: { $ref: '#/components/schemas/SystemUpdateStatus' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/system/update-runner/update:
    post:
      tags: [System]
      summary: Start guarded confirmed CMS update
      description: SystemAdmin/settings-management endpoint. Starts the configured host update script for a real update only when `CMS_UPDATE_RUNNER_ENABLED=true` and the request includes explicit confirmation. The API sets the configured confirmation environment variable before spawning the host script.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetRef: { type: [string, 'null'], description: Optional release, tag, branch, or commit ref to apply. }
                confirmed: { type: boolean, description: Must be true for real updates. }
                confirmation: { type: [string, 'null'], description: Alternative acknowledgement; `YES` is accepted. }
      responses:
        '202':
          description: Confirmed update request accepted.
          content:
            application/json:
              schema:
                type: object
                required: [updateStatus]
                properties:
                  updateStatus: { $ref: '#/components/schemas/SystemUpdateStatus' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/system/database-performance:
    get:
      tags: [System]
      summary: Get CMS database performance metrics
      description: SystemAdmin only. Returns a lightweight MySQL performance snapshot for the CMS database, including information_schema table sizing, selected server counters, and the largest tables by data plus index size. The endpoint uses the configured CMS database connection and returns a non-throwing snapshot with `connected=false` and `error` when metrics cannot be read.
      responses:
        '200':
          description: Database performance snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [databasePerformance]
                properties:
                  databasePerformance: { $ref: '#/components/schemas/DatabasePerformanceSnapshot' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/issue-types:
    get:
      tags: [System]
      summary: Get discovered issue type catalog
      description: SystemAdmin only. Returns the static issue category/severity/type catalog plus a current affected-instance list for each enabled issue type. The endpoint uses the configured CMS database connection and returns a non-throwing snapshot with `connected=false` and `error` when the catalog cannot be read.
      responses:
        '200':
          description: Issue type catalog snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [issueCatalog]
                properties:
                  issueCatalog: { $ref: '#/components/schemas/IssueCatalogSnapshot' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/poller/pause:
    post:
      tags: [System]
      summary: Pause the background polling runner
      description: SystemAdmin only. Pauses scheduled checks without stopping the API process.
      responses:
        '200':
          description: Poller paused.
          content:
            application/json:
              schema:
                type: object
                required: [poller]
                properties:
                  poller: { $ref: '#/components/schemas/InstancePollerStatus' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/poller/resume:
    post:
      tags: [System]
      summary: Resume the background polling runner
      description: SystemAdmin only. Resumes scheduled checks and recalculates next run time.
      responses:
        '200':
          description: Poller resumed.
          content:
            application/json:
              schema:
                type: object
                required: [poller]
                properties:
                  poller: { $ref: '#/components/schemas/InstancePollerStatus' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/system/poller/run-now:
    post:
      tags: [System]
      summary: Execute the background polling runner immediately
      description: SystemAdmin only. Forces an operator-initiated polling pass immediately, bypassing normal per-instance polling interval due checks while preserving in-flight instance protection.
      responses:
        '200':
          description: Poller run completed.
          content:
            application/json:
              schema:
                type: object
                required: [poller, summary]
                properties:
                  poller: { $ref: '#/components/schemas/InstancePollerStatus' }
                  summary:
                    oneOf:
                      - $ref: '#/components/schemas/InstancePollerSummary'
                      - type: 'null'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
...
    InstancePollerSummary:
      type: object
      required: [checked, skipped, failed]
      properties:
        checked: { type: integer, minimum: 0 }
        skipped: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
    InstancePollerStatus:
      type: object
      required: [state, isRunning, isPaused, tickIntervalMs, inFlight, lastRunAt, nextRunAt, lastSummary, lastError]
      properties:
        state: { type: string, enum: [running, paused, stopped] }
        isRunning: { type: boolean }
        isPaused: { type: boolean }
        tickIntervalMs: { type: integer, minimum: 1 }
        inFlight: { type: integer, minimum: 0 }
        lastRunAt: { type: [string, 'null'], format: date-time }
        nextRunAt: { type: [string, 'null'], format: date-time }
        lastSummary:
          oneOf:
            - $ref: '#/components/schemas/InstancePollerSummary'
            - type: 'null'
        lastError: { type: [string, 'null'] }
    QueueStatusSnapshot:
      type: object
      required: [enabled, mode, generatedAt, redis, bullBoard, queues]
      properties:
        enabled: { type: boolean }
        mode: { type: string, enum: [disabled, bullmq] }
        generatedAt: { type: string, format: date-time }
        redis:
          type: object
          required: [configured, connected, host, port, error]
          properties:
            configured: { type: boolean }
            connected: { type: boolean }
            host: { type: [string, 'null'] }
            port: { type: [integer, 'null'] }
            error: { type: [string, 'null'] }
        bullBoard:
          type: object
          required: [enabled, path]
          properties:
            enabled: { type: boolean }
            path: { type: [string, 'null'], description: Configured Bull Board mount path when enabled. Browser launch requires an authenticated API session; CMS Settings uses native queue visibility as the primary UI. }
        queues:
          type: array
          items: { $ref: '#/components/schemas/QueueStatusItem' }
    QueueStatusItem:
      type: object
      required: [name, description, waiting, active, delayed, failed, completed]
      properties:
        name: { type: string, enum: [instance-checks, database-maintenance, system-maintenance] }
        description: { type: string }
        waiting: { type: integer, minimum: 0 }
        active: { type: integer, minimum: 0 }
        delayed: { type: integer, minimum: 0 }
        failed:
          type: integer
          minimum: 0
          description: Retained failed job count reported by BullMQ; this is historical failed-job retention, not active work in progress.
        completed: { type: integer, minimum: 0 }
    QueueJobsSnapshot:
      type: object
      required: [enabled, mode, generatedAt, jobs]
      properties:
        enabled: { type: boolean }
        mode: { type: string, enum: [disabled, bullmq] }
        generatedAt: { type: string, format: date-time }
        jobs:
          type: array
          items: { $ref: '#/components/schemas/QueueJobSummary' }
    QueueJobSummary:
      type: object
      required: [id, queue, name, state, attemptsMade, timestamp, processedOn, finishedOn, failedReason, data]
      properties:
        id: { type: [string, 'null'] }
        queue: { type: string, enum: [instance-checks, database-maintenance, system-maintenance] }
        name: { type: string }
        state: { type: string, enum: [waiting, active, delayed, failed, completed, unknown] }
        attemptsMade: { type: integer, minimum: 0 }
        timestamp: { type: [string, 'null'], format: date-time }
        processedOn: { type: [string, 'null'], format: date-time }
        finishedOn: { type: [string, 'null'], format: date-time }
        failedReason: { type: [string, 'null'], description: Non-secret BullMQ failure summary. }
        data: { $ref: '#/components/schemas/QueueJobSafeData' }
...
    SystemUpdateStatus:
      type: object
      required: [generatedAt, runner, steps, lastRun, lastError]
      properties:
        generatedAt: { type: string, format: date-time }
        runner:
          type: object
          required: [enabled, state, inProgress, canRun, mode, command, dryRunCommand, requiresConfirmation, confirmationVariable, currentRef, targetRef]
          properties:
            enabled: { type: boolean }
            state: { type: string, enum: [idle, running, blocked, unavailable] }
            inProgress: { type: boolean }
            canRun: { type: boolean }
            mode: { type: string, enum: [host-script] }
            command: { type: string, example: scripts/deploy.sh update }
            dryRunCommand: { type: string, example: scripts/deploy.sh update --dry-run }
            requiresConfirmation: { type: boolean }
            confirmationVariable: { type: string, example: CONFIRM_UPDATE }
            currentRef: { type: [string, 'null'] }
            targetRef: { type: [string, 'null'] }
        steps:
          type: array
          items: { $ref: '#/components/schemas/SystemUpdateStep' }
        lastRun:
          oneOf:
            - $ref: '#/components/schemas/SystemUpdateRunSummary'
            - type: 'null'
        lastError: { type: [string, 'null'] }
...
    IssueCatalogSnapshot:
      type: object
      required: [configured, connected, generatedAt, error, categories, severities, issueTypes]
      properties:
        configured: { type: boolean }
        connected: { type: boolean }
        generatedAt: { type: string, format: date-time }
        error: { type: [string, 'null'] }
        categories:
          type: array
          items: { $ref: '#/components/schemas/IssueCategory' }
        severities:
          type: array
          items: { $ref: '#/components/schemas/IssueSeverity' }
        issueTypes:
          type: array
          items: { $ref: '#/components/schemas/DiscoveredIssueType' }
    IssueCategory:
      type: object
      required: [id, code, name, sortOrder]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        sortOrder: { type: integer }
...
    DatabasePerformanceSnapshot:
      type: object
      required: [configured, connected, database, generatedAt, error, schema, queryDigestStatus, summary, server, topTables, queryDigests]
      properties:
        configured: { type: boolean }
        connected: { type: boolean }
        database: { type: [string, 'null'] }
        generatedAt: { type: string, format: date-time }
        error: { type: [string, 'null'] }
        schema:
          $ref: '#/components/schemas/DatabaseSchemaPerformanceStatus'
        queryDigestStatus:
          $ref: '#/components/schemas/DatabaseQueryDigestStatus'
        summary:
          type: object
          required: [tableCount, estimatedRows, dataSizeBytes, indexSizeBytes, freeBytes, totalSizeBytes]
          properties:
            tableCount: { type: integer, minimum: 0 }
            estimatedRows: { type: integer, minimum: 0, description: Approximate row count from information_schema.tables.table_rows. }
            dataSizeBytes: { type: integer, minimum: 0 }
            indexSizeBytes: { type: integer, minimum: 0 }
            freeBytes: { type: integer, minimum: 0 }
            totalSizeBytes: { type: integer, minimum: 0 }
        server:
          type: object
          required: [version, uptimeSeconds, maxConnections, threadsConnected, maxUsedConnections, slowQueries, longQueryTimeSeconds, questions, abortedConnects, bufferPoolReadHitPercent]
          properties:
            version: { type: [string, 'null'] }
            uptimeSeconds: { type: [integer, 'null'], minimum: 0 }
            maxConnections: { type: [integer, 'null'], minimum: 0 }
            threadsConnected: { type: [integer, 'null'], minimum: 0 }
            maxUsedConnections: { type: [integer, 'null'], minimum: 0 }
            slowQueries: { type: [integer, 'null'], minimum: 0, description: MySQL `Slow_queries` status counter since server startup. }
            longQueryTimeSeconds: { type: [number, 'null'], minimum: 0, description: Current MySQL `long_query_time` threshold used to increment `Slow_queries`. }
            questions: { type: [integer, 'null'], minimum: 0 }
            abortedConnects: { type: [integer, 'null'], minimum: 0 }
            bufferPoolReadHitPercent: { type: [number, 'null'], minimum: 0, maximum: 100 }
        topTables:
          type: array
          items: { $ref: '#/components/schemas/DatabaseTablePerformance' }
        queryDigests:
          type: array
          description: Normalized statement patterns from performance_schema ordered by total execution time when available.
          items: { $ref: '#/components/schemas/DatabaseQueryDigestPerformance' }
...
    QueueScheduleSettings:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items: { $ref: '#/components/schemas/QueueScheduleJobSettings' }
...